refactor!: move CredentialSchema functions to KiltCredentialV1
