+++ packages/vc-export/src/index.ts
@@ -13,7 +13,6 @@ export * from './fromICredential.js'
 export * as KiltCredentialV1 from './KiltCredentialV1.js'
 export * as KiltAttestationProofV1 from './KiltAttestationProofV1.js'
 export * as KiltRevocationStatusV1 from './KiltRevocationStatusV1.js'
-export * as CredentialSchema from './CredentialSchema.js'
 export * as Presentation from './Presentation.js'
 export * as DidJWT from './DidJwt.js'
 export * as vcjs from './vc-js/index.js'
